Extract applying of newline style into own mod
